requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0
openpyxl>=3.1.0
pyarrow>=14.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
watchdog>=4.0.0
chardet>=5.2.0,<6.0.0
bcrypt>=4.1.0
statsmodels>=0.14.0
fastparquet>=2024.2.0
python-dotenv>=1.0.0
